ref: Moves the logic to parse ColibriStatsExtension
to Bridge, where it can be reused with the brewery detector.
